ad-hoc type casting warnings
This is an ugly patch that does ad-hoc type casting.  The correct
solution is to the refactor the code to not require such casting.

+++ src/ucm/util/replace.c
@@ -102,7 +102,7 @@ _UCM_DEFINE_DLSYM_FUNC(brk, ucm_orig_dlsym_brk, ucm_override_brk, int, -1, void*
 
 void *ucm_brk_syscall(void *addr)
 {
-    return (void*)syscall(SYS_brk, addr);
+    return (void*)(size_t) syscall(SYS_brk, addr);
 }
 
 int ucm_orig_brk(void *addr)
